Extract lookupFiles from utils to avoid node dependencies in browser bundle
