+++ lib/cli/collect-files.js
@@ -4,8 +4,8 @@ const path = require('path');
 const ansi = require('ansi-colors');
 const debug = require('debug')('mocha:cli:run:helpers');
 const minimatch = require('minimatch');
-const utils = require('../utils');
 const {NO_FILES_MATCH_PATTERN} = require('../errors').constants;
+const lookupFiles = require('./lookupFiles');
 
 /**
  * Exports a function that collects test files from CLI parameters.
@@ -27,7 +27,7 @@ module.exports = ({ignore, extension, file, recursive, sort, spec} = {}) => {
   spec.forEach(arg => {
     let newFiles;
     try {
-      newFiles = utils.lookupFiles(arg, extension, recursive);
+      newFiles = lookupFiles(arg, extension, recursive);
     } catch (err) {
       if (err.code === NO_FILES_MATCH_PATTERN) {
         unmatched.push({message: err.message, pattern: err.pattern});
